requirements.txt
amazon-textract-caller==0.2.4
amazon-textract-prettyprinter==0.1.10
amazon-textract-response-parser==0.1.48
boto3==1.35.6
botocore==1.35.6
jmespath==1.0.1
marshmallow==3.22.0
packaging==24.1
python-dateutil==2.9.0.post0
s3transfer==0.10.2
six==1.16.0
tabulate==0.9.0
urllib3==1.26.19
